requirements.txt
channels==3.0.4
channels_redis==4.0.0
daphne==3.0.2
Django==4.1.2
django-cloudinary-storage==0.3.0    
django-cors-headers==3.13.0
django-extensions==3.2.1
django-livereload-server==0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.4
python-dotenv==0.21.0
six==1.16.0
whitenoise==6.2.0